File needed as a workaround to an openssh config
Is not a real dotfile, but since I use jare/alpine image to build my ssh-server, and that base image has removed gendepends.sh from original alpine installation, I copied the content of a version of this file be able to run openssh. When I build my image I curl this file into the place it belongs.